+++ Engine/Source/Core/HAL/PlatformTime.h
@@ -11,7 +11,7 @@
 
 struct FPlatformTime
 {
-    static double Seconds();
+    static float Seconds();
     static uint64 Cycles64();
     static void   Sleep(float Seconds);
 };
